+++ Assets/ResonanceAudio/Scripts/ResonanceAudio.cs
@@ -72,8 +72,7 @@ public static class ResonanceAudio {
   /// @note This should only be called from the main Unity thread.
   public static void DisableRoomEffects() {
     // Set the room properties to null, which will effectively disable the room effects.
-    SetRoomProperties(IntPtr.Zero);
-    SetRt60ValuesAndProxyRoomProperties(null, IntPtr.Zero);
+    SetRoomProperties(IntPtr.Zero, null);
     if (roomPropertiesPtr != IntPtr.Zero) {
       // Free up the unmanaged memory.
       Marshal.FreeHGlobal(roomPropertiesPtr);
@@ -81,34 +80,29 @@ public static class ResonanceAudio {
     }
   }
 
-  /// Updates the room effects of the environment with given |room| properties.
+  /// Updates the room effects of the environment with the given |room|.
   /// @note This should only be called from the main Unity thread.
-  public static void UpdateRoom(ResonanceAudioRoom currentRoom) {
+  public static void UpdateRoom(ResonanceAudioRoom room) {
     if (roomPropertiesPtr == IntPtr.Zero) {
       // Allocate the unmanaged memory only once.
       roomPropertiesPtr = Marshal.AllocHGlobal(Marshal.SizeOf(roomProperties));
     }
-    UpdateRoomProperties(currentRoom);
+    UpdateRoomProperties(room);
     Marshal.StructureToPtr(roomProperties, roomPropertiesPtr, false);
-    SetRoomProperties(roomPropertiesPtr);
-    SetRt60ValuesAndProxyRoomProperties(null, IntPtr.Zero);
+    SetRoomProperties(roomPropertiesPtr, null);
     Marshal.DestroyStructure(roomPropertiesPtr, typeof(RoomProperties));
   }
 
   /// Updates the room effects of the environment with given |reverbProbe|.
   /// @note This should only be called from the main Unity thread.
-  public static void UpdateReverbProbe(ResonanceAudioReverbProbe currentReverbProbe) {
+  public static void UpdateReverbProbe(ResonanceAudioReverbProbe reverbPobe) {
     if (roomPropertiesPtr == IntPtr.Zero) {
       // Allocate the unmanaged memory only once.
       roomPropertiesPtr = Marshal.AllocHGlobal(Marshal.SizeOf(roomProperties));
     }
-    UpdateRoomProperties(currentReverbProbe);
+    UpdateRoomProperties(reverbPobe);
     Marshal.StructureToPtr(roomProperties, roomPropertiesPtr, false);
-    SetRoomProperties(roomPropertiesPtr);
-    // Proxy room properties. Used in calculating reflection effects.
-    UpdateProxyRoomProperties(currentReverbProbe);
-    Marshal.StructureToPtr(roomProperties, roomPropertiesPtr, true);
-    SetRt60ValuesAndProxyRoomProperties(currentReverbProbe.rt60s, roomPropertiesPtr);
+    SetRoomProperties(roomPropertiesPtr, reverbPobe.rt60s);
     Marshal.DestroyStructure(roomPropertiesPtr, typeof(RoomProperties));
   }
 
@@ -144,27 +138,36 @@ public static class ResonanceAudio {
 
   /// Computes the RT60s and proxy room properties.
   /// @note This should only be called from the main Unity thread.
-  public static bool ComputeRt60sAndProxyRoom(
-    int totalNumPaths, int numPathsPerBatch, int maxDepth, float energyThreshold,
-    Vector3 probePosition, float listenerSphereRadius, ref float[] outputRt60s,
-    ref RoomProperties outputProxyRoomProperties) {
+  public static bool ComputeRt60sAndProxyRoom(ResonanceAudioReverbProbe reverbProbe,
+                                              int totalNumPaths, int numPathsPerBatch, int maxDepth,
+                                              float energyThreshold, float listenerSphereRadius) {
 #if UNITY_EDITOR
-    float[] probePositionFloatArray =
-      new float[] {probePosition.x, probePosition.y, probePosition.z};
-    IntPtr outputProxyRoomPropertiesPtr =
+    Vector3 reverbProbePosition = reverbProbe.transform.position;
+    roomPosition[0] = reverbProbePosition.x;
+    roomPosition[1] = reverbProbePosition.y;
+    roomPosition[2] = reverbProbePosition.z;
+    IntPtr proxyRoomPropertiesPtr =
       Marshal.AllocHGlobal(Marshal.SizeOf(typeof(RoomProperties)));
     float sampleRate = (float) AudioSettings.GetConfiguration().sampleRate;
     int impulseResponseNumSamples = (int) (maxReverbTime * sampleRate);
     if (!ComputeRt60sAndProxyRoom(totalNumPaths, numPathsPerBatch, maxDepth, energyThreshold,
-                                  probePositionFloatArray, listenerSphereRadius, sampleRate,
-                                  impulseResponseNumSamples, outputRt60s,
-                                  outputProxyRoomPropertiesPtr)) {
+                                  roomPosition, listenerSphereRadius, sampleRate,
+                                  impulseResponseNumSamples, reverbProbe.rt60s,
+                                  proxyRoomPropertiesPtr)) {
       return false;
     }
 
-    outputProxyRoomProperties = (RoomProperties) Marshal.PtrToStructure(
-      outputProxyRoomPropertiesPtr, typeof(RoomProperties));
-    Marshal.FreeHGlobal(outputProxyRoomPropertiesPtr);
+    // Validate the computed RT60s for all frequency bands to the reverb probe.
+    for (int band = 0; band < reverbProbe.rt60s.Length; ++band) {
+        reverbProbe.rt60s[band] = Mathf.Clamp(reverbProbe.rt60s[band], 0.0f,
+                                              ResonanceAudio.maxReverbTime);
+    }
+    // Copy the estimated proxy room properties to the reverb probe.
+    RoomProperties proxyRoomProperties =
+        (RoomProperties) Marshal.PtrToStructure(proxyRoomPropertiesPtr, typeof(RoomProperties));
+    SetProxyRoomProperties(reverbProbe, proxyRoomProperties);
+    Marshal.FreeHGlobal(proxyRoomPropertiesPtr);
+
     return true;
 #else
     return false;
@@ -252,7 +255,7 @@ public static class ResonanceAudio {
   public const float occlusionDetectionInterval = 0.2f;
 
   [StructLayout(LayoutKind.Sequential)]
-  public class RoomProperties {
+  private class RoomProperties {
     // Center position of the room in world space.
     public float positionX;
     public float positionY;
@@ -301,6 +304,28 @@ public static class ResonanceAudio {
     rotation = Quaternion.LookRotation(transformMatrix.GetColumn(2), transformMatrix.GetColumn(1));
   }
 
+  /// Sets computed |proxyRoomProperties| for the given |reverbProbe|. Proxy rooms are estimated by
+  /// the ray-tracing engine and passed back to be used in real-time early reflections.
+  private static void SetProxyRoomProperties(ResonanceAudioReverbProbe reverbProbe,
+                                             RoomProperties proxyRoomProperties) {
+    reverbProbe.proxyRoomPosition.x = proxyRoomProperties.positionX;
+    reverbProbe.proxyRoomPosition.y = proxyRoomProperties.positionY;
+    reverbProbe.proxyRoomPosition.z = proxyRoomProperties.positionZ;
+    reverbProbe.proxyRoomRotation.x = proxyRoomProperties.rotationX;
+    reverbProbe.proxyRoomRotation.y = proxyRoomProperties.rotationY;
+    reverbProbe.proxyRoomRotation.z = proxyRoomProperties.rotationZ;
+    reverbProbe.proxyRoomRotation.w = proxyRoomProperties.rotationW;
+    reverbProbe.proxyRoomSize.x = proxyRoomProperties.dimensionsX;
+    reverbProbe.proxyRoomSize.y = proxyRoomProperties.dimensionsY;
+    reverbProbe.proxyRoomSize.z = proxyRoomProperties.dimensionsZ;
+    reverbProbe.proxyRoomLeftWall = proxyRoomProperties.materialLeft;
+    reverbProbe.proxyRoomRightWall = proxyRoomProperties.materialRight;
+    reverbProbe.proxyRoomFloor = proxyRoomProperties.materialBottom;
+    reverbProbe.proxyRoomCeiling = proxyRoomProperties.materialTop;
+    reverbProbe.proxyRoomBackWall = proxyRoomProperties.materialBack;
+    reverbProbe.proxyRoomFrontWall = proxyRoomProperties.materialFront;
+  }
+
   // Updates room properties with the given |room|.
   private static void UpdateRoomProperties(ResonanceAudioRoom room) {
     FillGeometryOfRoomProperties(room.transform.position, room.transform.rotation,
@@ -311,35 +336,10 @@ public static class ResonanceAudio {
                                   room.reflectivity);
   }
 
-  // Updates room properties with the given |reverbProbe|. For a reverb probe, the room properties
-  // are used only in computing the distance attenuation.
+  // Updates room properties with the given |reverbProbe|.
   private static void UpdateRoomProperties(ResonanceAudioReverbProbe reverbProbe) {
-    Vector3 scale = Vector3.zero;
-    if (reverbProbe.runtimeApplicationRegionShape ==
-        ResonanceAudioReverbProbe.ApplicationRegionShape.Sphere) {
-      // Use the minimum enclosing cube of the sphere for distance attenuation.
-      var diameter = 2.0f * reverbProbe.GetScaledSphericalApplicationRegionRadius();
-      scale = diameter * Vector3.one;
-    } else {
-      scale = reverbProbe.GetScaledBoxApplicationRegionSize();
-    }
-    FillGeometryOfRoomProperties(reverbProbe.transform.position, reverbProbe.transform.rotation,
-                                 scale);
-    // Surface materials are not needed for reverb probes.
-    var surfaceMaterial = ResonanceAudioRoomManager.SurfaceMaterial.Transparent;
-    FillWallMaterialsOfRoomProperties(surfaceMaterial, surfaceMaterial, surfaceMaterial,
-                                      surfaceMaterial, surfaceMaterial, surfaceMaterial);
-    // Reflectivity is not a responsibility of reverb probes.
-    float reflectivity = 0.0f;
-    FillModifiersOfRoomProperties(reverbProbe.reverbGainDb, reverbProbe.reverbTime,
-                                  reverbProbe.reverbBrightness, reflectivity);
-  }
-
-  // Updates room properties with the proxy room of the given |reverbProbe|.
-  private static void UpdateProxyRoomProperties(ResonanceAudioReverbProbe reverbProbe) {
     FillGeometryOfRoomProperties(reverbProbe.proxyRoomPosition, reverbProbe.proxyRoomRotation,
                                  reverbProbe.proxyRoomSize);
-
     FillWallMaterialsOfRoomProperties(reverbProbe.proxyRoomLeftWall, reverbProbe.proxyRoomRightWall,
                                       reverbProbe.proxyRoomFloor, reverbProbe.proxyRoomCeiling,
                                       reverbProbe.proxyRoomFrontWall,
@@ -401,6 +401,9 @@ public static class ResonanceAudio {
   // Occlusion layer mask.
   private static int occlusionMaskValue = -1;
 
+  // Pre-allocated position array for proxy room computation.
+  private static float[] roomPosition = new float[3];
+
   // Pre-allocated room properties instance for room effects computation.
   private static RoomProperties roomProperties = new RoomProperties();
 
@@ -425,11 +428,7 @@ public static class ResonanceAudio {
 
   // Room handlers.
   [DllImport(pluginName)]
-  private static extern void SetRoomProperties(IntPtr roomProperties);
-
-  [DllImport(pluginName)]
-  private static extern void SetRt60ValuesAndProxyRoomProperties(float[] rt60s,
-                                                                 IntPtr proxyRoomProperties);
+  private static extern void SetRoomProperties(IntPtr roomProperties, float[] rt60s);
 
 #if UNITY_EDITOR
   // Soundfield recorder handlers.