+++ Assets/ResonanceAudio/Scripts/ResonanceAudioMaterialMapper.cs
@@ -168,7 +168,6 @@ public class ResonanceAudioMaterialMapper : ScriptableObject {
       var acousticMesh = ResonanceAudioAcousticMesh.GenerateFromMeshFilter(
           gameObject.GetComponent<MeshFilter>(), surfaceMaterialShader);
       if (acousticMesh == null) {
-        Debug.LogError("acousticMesh == null");
         continue;
       }