+++ Assets/ResonanceAudio/Scripts/ResonanceAudioReverbProbe.cs
@@ -23,7 +23,7 @@ using System.Collections;
 [ExecuteInEditMode]
 public class ResonanceAudioReverbProbe : MonoBehaviour {
   /// Supported choices of the shape of regions of application.
-  public enum ApplicationRegionShape {
+  public enum RegionShape {
     Sphere = 0,
     Box = 1
   }
@@ -49,18 +49,18 @@ public class ResonanceAudioReverbProbe : MonoBehaviour {
   /// Which shape of regions of application to use at runtime to check whether the listener is
   /// inside.
   [Tooltip("Shape of the region of application of this reverb.")]
-  public ApplicationRegionShape runtimeApplicationRegionShape = ApplicationRegionShape.Box;
+  public RegionShape regionShape = RegionShape.Box;
 
   /// Size of the box-shaped region of application, normalized with respect to scale of the game
   /// object.
   [Tooltip("Sets the dimensions of a box-shaped region of application in meters relative to the " +
            "scale of the game object.")]
-  public Vector3 boxApplicationRegionSize = Vector3.one;
+  public Vector3 boxRegionSize = Vector3.one;
 
   /// Radius of the sphere-shaped region of application.
   [Tooltip("Sets the radius of a spherical region of application in meters relative to the scale " +
            "of the game object.")]
-  public float sphereApplicationRegionRadius = 1.0f;
+  public float sphereRegionRadius = 1.0f;
 
   /// Only apply the reverb properties if this probe is visible to the listener.
   [Tooltip("Applies this reverb only when the center of the probe is visible from the listener. " +
@@ -105,53 +105,30 @@ public class ResonanceAudioReverbProbe : MonoBehaviour {
     ResonanceAudioRoomManager.UpdateReverbProbe(this);
   }
 
-  /// Draw either a box or sphere wire-frame depending on the runtime shape of the region of
-  /// application.
-  void OnDrawGizmosSelected() {
-    Gizmos.color = Color.magenta;
-    switch (runtimeApplicationRegionShape) {
-    case ApplicationRegionShape.Sphere:
-      Gizmos.DrawWireSphere(transform.position, GetScaledSphericalApplicationRegionRadius());
-      break;
-    case ApplicationRegionShape.Box:
-      Gizmos.matrix = transform.localToWorldMatrix;
-      Gizmos.DrawWireCube(Vector3.zero, boxApplicationRegionSize);
-      break;
-    }
-  }
-
-  /// Set proxy room properties. Proxy rooms are estimated by the ray-tracing engine and passed
-  /// back to be used in real-time early reflections.
-  public void SetProxyRoomProperties(ResonanceAudio.RoomProperties proxyRoomProperties) {
-    proxyRoomPosition = new Vector3(proxyRoomProperties.positionX,
-                                    proxyRoomProperties.positionY,
-                                    proxyRoomProperties.positionZ);
-    proxyRoomRotation = new Quaternion(proxyRoomProperties.rotationX,
-                                       proxyRoomProperties.rotationY,
-                                       proxyRoomProperties.rotationZ,
-                                       proxyRoomProperties.rotationW);
-    proxyRoomSize = new Vector3(proxyRoomProperties.dimensionsX,
-                                proxyRoomProperties.dimensionsY,
-                                proxyRoomProperties.dimensionsZ);
-    proxyRoomLeftWall = proxyRoomProperties.materialLeft;
-    proxyRoomRightWall = proxyRoomProperties.materialRight;
-    proxyRoomFloor = proxyRoomProperties.materialBottom;
-    proxyRoomCeiling = proxyRoomProperties.materialTop;
-    proxyRoomBackWall = proxyRoomProperties.materialBack;
-    proxyRoomFrontWall = proxyRoomProperties.materialFront;
-  }
-
   /// Gets the radius of the spherical region of application scaled by the transform. In order to
   /// maintain the spherical shape, the maximum of the scales in three dimensions is used to
   /// scale the radius (similar to how Unity handles Sphere Collider).
-  public float GetScaledSphericalApplicationRegionRadius() {
+  public float GetScaledSphericalRegionRadius() {
     Vector3 scale = transform.lossyScale;
     float maxScale = Mathf.Max(Mathf.Max(scale.x, scale.y), scale.z);
-    return sphereApplicationRegionRadius * maxScale;
+    return sphereRegionRadius * maxScale;
   }
 
   /// Gets the size of the box-shaped region of application scaled by the transform.
-  public Vector3 GetScaledBoxApplicationRegionSize() {
-    return Vector3.Scale(transform.lossyScale, boxApplicationRegionSize);
+  public Vector3 GetScaledBoxRegionSize() {
+    return Vector3.Scale(transform.lossyScale, boxRegionSize);
+  }
+
+  void OnDrawGizmosSelected() {
+    Gizmos.color = Color.magenta;
+    switch (regionShape) {
+    case RegionShape.Sphere:
+      Gizmos.DrawWireSphere(transform.position, GetScaledSphericalRegionRadius());
+      break;
+    case RegionShape.Box:
+      Gizmos.matrix = transform.localToWorldMatrix;
+      Gizmos.DrawWireCube(Vector3.zero, boxRegionSize);
+      break;
+    }
   }
 }